🎨Reforming : 엔티티 id 값 Long 타입으로 변경

+++ src/main/java/practice/single1_11war/domain/cart/repository/CartProductRepository.java
@@ -3,5 +3,5 @@ package practice.single1_11war.domain.cart.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import practice.single1_11war.domain.cart.entity.CartProduct;
 
-public interface CartProductRepository extends JpaRepository<CartProduct, Integer> {
+public interface CartProductRepository extends JpaRepository<CartProduct, Long> {
 }
